Refactor `Response` and `ResponseEmitter` classes into `Engine\Response` namespace and update references
