+++ Engine/JsonResponse.php
@@ -2,6 +2,7 @@
 
 namespace Engine;
 
+use Engine\Response\Response;
 use JsonException;
 
 class JsonResponse extends Response
